brain/pyproject.toml
[project]
name = "brain"
version = "0.1.0"
description = "LLMWay Brain — internal provider/model management module"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "redis[asyncio]>=5.0",
    "sqlalchemy[asyncio]>=2.0",
    "shared",
]

[tool.uv.sources]
shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/brain"]
